refactor(reservation): Payment, Reservation 요청 DTO 통일
