+++ src/main/java/roomescape/payment/application/service/PaymentClient.java
@@ -1,9 +1,9 @@
 package roomescape.payment.application.service;
 
 import roomescape.payment.domain.Payment;
-import roomescape.payment.presentation.dto.PaymentRequest;
+import roomescape.reservation.presentation.dto.ReservationRequest;
 
 public interface PaymentClient {
 
-    Payment approve(final PaymentRequest paymentRequest);
+    Payment approve(final ReservationRequest reservationRequest);
 }
