+++ src/main/java/roomescape/reservation/presentation/controller/ReservationController.java
@@ -17,9 +17,8 @@ import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import roomescape.global.auth.Auth;
 import roomescape.member.domain.Role;
 import roomescape.member.presentation.resolver.LoginMember;
-import roomescape.payment.presentation.dto.PaymentRequest;
-import roomescape.reservation.application.service.ReservationService;
 import roomescape.reservation.presentation.dto.ReservationRequest;
+import roomescape.reservation.application.service.ReservationService;
 import roomescape.reservation.presentation.dto.ReservationResponse;
 import roomescape.reservation.presentation.dto.UserReservationsResponse;
 
@@ -36,10 +35,10 @@ public class ReservationController {
     @Auth(Role.USER)
     @PostMapping
     public ResponseEntity<ReservationResponse> createReservation(
-            final @RequestBody @Valid PaymentRequest paymentRequest,
+            final @RequestBody @Valid ReservationRequest reservationRequest,
             final @LoginMember Long memberId
     ) {
-        ReservationResponse reservation = reservationService.createUserReservationAndPayment(paymentRequest, memberId);
+        ReservationResponse reservation = reservationService.createUserReservationAndPayment(reservationRequest, memberId);
 
         return ResponseEntity.created(createUri(reservation.getId()))
                 .body(reservation);
